eevee_next/requirements.txt
fake-bpy-module-4.2